+++ test/operation/params_test.rb
@@ -7,6 +7,7 @@ class Trailblazer::Operation
     end
   end
 
+  # Returned object will replace "params". Original is saved in "params.original".
   Params::Replace = ->(input, options) {
     options[:skills]["params.original"] = original = options[:skills]["params"]
      options[:skills]["params"] = input.params!(original)